requirements.txt
APScheduler==3.5.3
bokeh==1.0.4
configparser==3.7.3
Django==2.1.7
django-rest-framework==0.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
et-xmlfile==1.0.1
grpcio==1.18.0
grpcio-tools==1.18.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.0
netifaces==0.10.9
numpy==1.16.1
openpyxl==2.6.0
packaging==19.0
pandas==0.24.1
Pillow==5.4.1
protobuf==3.6.1
psutil==5.4.8
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
six==1.12.0
tornado==5.1.1
tzlocal==1.5.1
wheel==0.33.1
